removed unused code and improved dmg indicator and fake lag

+++ DarkToolX/core/features/backtrack.cpp
@@ -80,7 +80,7 @@ void features::backtrack::update()
 
 		records[i].push_front(rec);
 
-		while (records[i].size() > 3 && records[i].size() > static_cast<size_t>(time_to_ticks(static_cast<float>(csgo::conf->aimbot().backtrack_time_limit) / 1000.f)))
+		while (records[i].size() > 3 && records[i].size() > static_cast<size_t>(time_to_ticks(csgo::conf->aimbot().backtrack_time_limit / 1000.f)))
 			records[i].pop_back();
 
 		if (auto invalid = std::find_if(std::cbegin(records[i]), std::cend(records[i]), [](const record& rec) { return !valid(rec.simulation_time); }); invalid != std::cend(records[i]))
